Módulo Central/django-aula-master/requirements.txt
Django==1.9.5
django-classy-tags==0.7.2
django-cms==3.2.3
django-cors-headers==1.1.0
django-crispy-forms==1.6.0
django-data-exports==0.7
django-extensions==1.6.1
django-formtools==1.0
django-inspect-model==0.7
django-sekizai==0.9.0
django-treebeard==4.0
djangocms-admin-style==1.1.0
graphviz==0.4.10
html5lib==0.9999999
mysqlclient==1.3.7
Pillow==3.2.0
pydot==1.0.2
pyparsing==1.5.7
six==1.10.0
